+++ CHANGELOG.md
@@ -2,7 +2,39 @@
 
 New features, fixed bugs, known defects and other noteworthy changes to each release of the Catena-X Portal Backend.
 
-## Unreleased
+## 1.0.0-RC7
+
+### Changen
+n/a
+
+### Feature
+n/a
+
+### Technical Support
+* enable provisioning, appmarketplace and services service to use config from env vars
+
+### Bugfix
+* double creation of notifications for app activation fixed
+* registration approval: removing the company user assigned roles for a specific client id while activating the company
+
+## 1.0.0-RC6
+
+### Change
+* Apps & Service Services: auto set the releaseDate of an offer with the approval endpoint
+* App Service:
+  * POST endpoint to upload documents for services in currently under creation
+  * Endpoint controller switch of /app/decline from apps to appReleaseProcess ![Tag](https://img.shields.io/static/v1?label=&message=BreakingChange&color=yellow&style=flat)
+* Checklist-Worker for Registration Service: updates implemented (full feature delivery still ongoing)
+
+### Feature
+n/a
+
+### Technical Support
+* enable migrations, portal maintenance and checklist-worker jobs as well as notification service to use config from env vars
+* remove initdb container setup - replaced by configmap in portal helm chart
+
+### Bugfix
+n/a
 
 ## 1.0.0-RC5